+++ R/epw.R
@@ -1505,7 +1505,7 @@ format.Epw <- function (x, ...) {
 download_weather <- function (pattern, filename = NULL, dir = ".", type = c("all", "epw", "ddy"),
                               ask = TRUE, max_match = 3) {
     pattern <- gsub("\\s+", ".", pattern)
-    d <- as.data.table(weather_db)
+    d <- as.data.table(WEATHER_DB)
     res <- d[stri_detect_regex(title, pattern, case_insensitive = TRUE)]
 
     mes_location <- function (index = NULL, title, country, state_province, location, wmo_number, source_type, longitude, latitude) {