+++ tests/testthat/test_epw.R
@@ -247,7 +247,8 @@ test_that("Epw class", {
     # do not test on CRAN
     skip_on_cran()
     # download weather
-    expect_silent({path_epw <- download_weather("USA_CA_San.Francisco.Intl.AP.724940_TMY3",
+    eplusr_option(verbose_info = TRUE)
+    expect_message({path_epw <- download_weather("USA_CA_San.Francisco.Intl.AP.724940_TMY3",
         ask = FALSE, type = "epw", dir = tempdir())}
     )
 })